feat: Add release signing configuration and keystore properties.
This commit introduces a release signing configuration using a `keystore.properties` file. It loads the keystore properties from this file and assigns them to the `signingConfigs` block in `build.gradle.kts`.  Also, the release build type is configured to use this release configuration, and the keystore.properties and the .jks keystore file is added to the .gitignore files. Finally, debug is configured with the release configuration as well.