Refactor: Update deployment target selection in `.idea/deploymentTargetSelector.xml`.
This commit updates the selected device and timestamp in the `DropdownSelection` within the `.idea/deploymentTargetSelector.xml` file. The selected target has been switched from an emulator to a physical device, and the timestamp has been updated from `2025-01-21T05:33:18.661401600Z` to `2025-01-21T06:31:28.382724500Z`.
This file is managed by the IDE, and changes can be ignored.