refactor: Reorder dialog and bottom sheet parameters
This commit reorders the parameters of the `DeletePersonDialog`, `InviteBottomSheet`, and `LeaveCourseDialog` composables for consistency. Specifically, the `onDismissRequest` lambda is moved to be the last parameter in these components. This refactoring improves code readability and consistency across different UI components.
It also moves the `show` parameter before `onDismissRequest` in `InviteBottomSheet` composable.
- `LeaveCourseDialog` `onDismissRequest` is moved to be the last parameter.
- `DeletePersonDialog` `onDismissRequest` is moved to be the last parameter.
- `InviteBottomSheet` `show` parameter is moved before `onDismissRequest` .