feat: Add IME padding to Course Details screen and EdumateNavHost
This commit adds `imePadding` to the Course Details screen content to avoid content being obscured by the keyboard.
`imePadding` is also applied to EdumateNavHost to ensure proper handling of the IME.
Also removes unnecessary `imePadding` modifier in MainActivity.
This change is made to avoid adding the navigation bar padding twice, and solve the `adjustResize` soft keyboard issue on some devices.
`consumeWindowInsets` is used in conjunction with `imePadding` to avoid unexpected padding issues.