feat: Display Lottie animation conditionally in error screen
This commit modifies the `AnimatedErrorScreen` to conditionally display the Lottie animation based on available space. The Lottie animation and the spacer will only be shown if the available space height is greater than 260.dp. This improves the layout on smaller screens.
Additionally the commit introduces `BoxWithConstraints` to determine screen dimensions.